In print_extensions(), declare the *_idx variables as int instead of size_t.
While it shouldn't make a difference, it makes more sense to me.
It's also consistent with (at least) print_crl().

+++ lib/x509/output.c
@@ -691,14 +691,14 @@ print_extensions (gnutls_string * str, const char *prefix, int type,
 		  cert_type_t cert)
 {
   int i, err;
-  size_t san_idx = 0;
-  size_t proxy_idx = 0;
-  size_t basic_idx = 0;
-  size_t keyusage_idx = 0;
-  size_t keypurpose_idx = 0;
-  size_t ski_idx = 0;
-  size_t aki_idx = 0;
-  size_t crldist_idx = 0;
+  int san_idx = 0;
+  int proxy_idx = 0;
+  int basic_idx = 0;
+  int keyusage_idx = 0;
+  int keypurpose_idx = 0;
+  int ski_idx = 0;
+  int aki_idx = 0;
+  int crldist_idx = 0;
 
   for (i = 0;; i++)
     {
